requirement.txt
aliyun-python-sdk==2.2.0
aliyun-python-sdk-alidns==2.0.14
aliyun-python-sdk-core==2.13.13
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-domain==3.14.2
aliyun-python-sdk-kms==2.10.1
AxmlParserPY==0.0.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
crcmod==1.7
cryptography==2.8
Django==1.11.14
django-cors-headers==3.2.1
django-mirage-field==1.0.1
get==2019.4.13
idna==2.8
jmespath==0.9.4
oss2==2.9.1
post==2019.4.13
public==2019.4.13
pycparser==2.20
pycryptodome==3.9.7
PyMySQL==0.9.3
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.14.0
sqlparse==0.3.1
tqdm==4.43.0
urllib3==1.25.7
